🏗️ UI를 별도 모듈로 분리.
`:app` 모듈에는 애플리케이션으로 실행할 때 필요한 프로세스 관리(`Application`), UI 구동 영역 확보(`Activity`), 애플리케이션 컴포넌트 생성(Hilt 컴포넌트 IoC)만 남긴다.
